requirements.txt
git+https://github.com/elliottech/lighter-python.git
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp>=3.11.0
websockets==12.0
requests>=2.31.0
